CVStreamlineApp-main/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
argilla==1.10.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
blinker==1.7.0
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.2.0
click==8.1.4
commonmark==0.9.1
cryptography==41.0.2
customtkinter==5.0.3
darkdetect==0.8.0
dataclasses-json==0.5.9
Deprecated==1.2.14
et-xmlfile==1.1.0
filetype==1.2.0
Flask==3.0.0
frozenlist==1.3.3
h11==0.14.0
holidays==0.40
httpcore==0.16.3
httpx==0.23.3
idna==3.4
inexactsearch==1.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
langchain==0.0.229
langchainplus-sdk==0.0.20
lxml==4.9.3
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.1
pandas==1.5.3
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.4.0
pycparser==2.21
pydantic==1.10.11
Pygments==2.15.1
pypandoc==1.11
pyspellchecker==0.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.0.1
silpa-common==0.3
six==1.16.0
sniffio==1.3.0
soundex==1.1.3
spellchecker==0.4
SQLAlchemy==2.0.18
tabulate==0.9.0
tenacity==8.2.2
tqdm==4.65.0
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.7.1
unstructured==0.8.0
urllib3==2.0.3
Werkzeug==3.0.1
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
